Crate x25519_xchacha20poly1305

Authenticated Public Key Encryption using XChaChaPoly20Poly1305 with an X25519 key exchange.

Warning

This crate is written by someone who is not a cryptographer. And, while the dependencies of this crate are excellent crates, most of the code that this crate is built upon has not been audited. Use at your own risk.

Public Key Encryption

This crate uses X25519 for public key encryption between two parties. Each party has their own StaticSecret. A PublicKey can be derived from each secret and can be freely exchanged. Using PublicKeyEncryption, payloads can be encrypted and decrypted by either party as long as both parties know each other’s public keys.

use x25519_xchacha20poly1305::{PublicKeyEncryption, x25519::{PublicKey, StaticSecret}};
use rand::{thread_rng, Rng};

let bob_private_key = StaticSecret::new(thread_rng());
let bob_public_key = PublicKey::from(&bob_private_key);

let mary_private_key = StaticSecret::new(thread_rng());
let mary_public_key = PublicKey::from(&mary_private_key);

// Mary and Bob can exchange public keys, which will allow them to encrypt and decrypt
// messages for each other.
let bob = PublicKeyEncryption::new(&bob_private_key, &mary_public_key);
let mary = PublicKeyEncryption::new(&mary_private_key, &bob_public_key);

// Bob sends a message to Mary with a random nonce.
let nonce: [u8; 24] = thread_rng().gen();
let bobs_message = bob.encrypt(b"hello", b"associated data", &nonce).unwrap();

// Bob can decrypt this message.
assert_eq!(
    bob.decrypt(&bobs_message, b"associated data", &nonce).unwrap(),
    b"hello"
);

// And, so can Mary.
assert_eq!(
    mary.decrypt(&bobs_message, b"associated data", &nonce).unwrap(),
    b"hello"
);

Re-exports

pub use x25519_dalek as x25519;

Modules

ephemeral
Encrypts messages for a StaticSecret using only the PublicKey.

Structs

PublicKeyEncryption
A public key encryption context.

Traits

Secret
A secret that can be used with PublicKeyEncryption.