Module x25519_xchacha20poly1305::ephemeral

Encrypts messages for a StaticSecret using only the PublicKey.

This is accomplished by generating an EphemeralSecret. The ephemeral secret key + the known public key are used to encrypt the payload. The ephemeral public key is embedded in the resulting payload. The ephemeral secret key is forgotten after encrypting the payload.

The order of the produced payload is:

Ephemeral X25519 Public Key: 32 bytes

Poly1305 MAC: 16 bytes

Remainder: ciphertext

This is an example showing how to encrypt a message using a public key and decrypt it using the private key:

use x25519_xchacha20poly1305::{
    ephemeral::{PublicKeyExt as _, StaticSecretExt as _},
    x25519::{PublicKey, StaticSecret},
};
use rand::{thread_rng, Rng};

let private_key = StaticSecret::new(thread_rng());
let public_key = PublicKey::from(&private_key);

let nonce: [u8; 24] = thread_rng().gen();
let encrypted = public_key.encrypt(b"hello", b"associated data", &nonce).unwrap();
let decrypted = private_key.decrypt(&encrypted, b"associated data", &nonce).unwrap();
assert_eq!(decrypted, b"hello");

Traits

PublicKeyExt
Extension trait for PublicKey. Encrypts plaintext using XChaCha20Poly1305 that only the owner of the private key of self can decrypt. For more information on how this is accomplished, see the ephemeral module.

StaticSecretExt
Extension trait for StaticSecret. Decrypts paylods produced by PublicKeyExt::encrypt(). For more information on how this is accomplished, see the ephemeral module.