Trait x25519_xchacha20poly1305::ephemeral::StaticSecretExt

pub trait StaticSecretExt {
    fn decrypt(
        &self, 
        payload: &[u8], 
        associated_data: &[u8], 
        nonce: &[u8]
    ) -> Result<Vec<u8>, Error>;
}

Extension trait for StaticSecret. Decrypts paylods produced by PublicKeyExt::encrypt(). For more information on how this is accomplished, see the ephemeral module.

Required methods

fn decrypt(
&self,
payload: &[u8],
associated_data: &[u8],
nonce: &[u8]
) -> Result<Vec<u8>, Error>

Decrypts a payload that was previously encrypted with PublicKeyExt. For more information on how this is accomplished, see the ephemeral module.

Loading content...